src/main/webapp/app/entities/payment-file-header/delete/payment-file-header-delete-dialog.component.spec.ts
jest.mock('@ng-bootstrap/ng-bootstrap');

import { ComponentFixture, TestBed, inject, fakeAsync, tick } from '@angular/core/testing';
import { HttpResponse } from '@angular/common/http';
import { HttpClientTestingModule } from '@angular/common/http/testing';
import { of } from 'rxjs';
import { NgbActiveModal } from '@ng-bootstrap/ng-bootstrap';

import { PaymentFileHeaderService } from '../service/payment-file-header.service';

import { PaymentFileHeaderDeleteDialogComponent } from './payment-file-header-delete-dialog.component';

describe('PaymentFileHeader Management Delete Component', () => {
  let comp: PaymentFileHeaderDeleteDialogComponent;
  let fixture: ComponentFixture<PaymentFileHeaderDeleteDialogComponent>;
  let service: PaymentFileHeaderService;
  let mockActiveModal: NgbActiveModal;

  beforeEach(() => {
    TestBed.configureTestingModule({
      imports: [HttpClientTestingModule],
      declarations: [PaymentFileHeaderDeleteDialogComponent],
      providers: [NgbActiveModal],
    })
      .overrideTemplate(PaymentFileHeaderDeleteDialogComponent, '')
      .compileComponents();
    fixture = TestBed.createComponent(PaymentFileHeaderDeleteDialogComponent);
    comp = fixture.componentInstance;
    service = TestBed.inject(PaymentFileHeaderService);
    mockActiveModal = TestBed.inject(NgbActiveModal);
  });

  describe('confirmDelete', () => {
    it('Should call delete service on confirmDelete', inject(
      [],
      fakeAsync(() => {
        // GIVEN
        jest.spyOn(service, 'delete').mockReturnValue(of(new HttpResponse({ body: {} })));

        // WHEN
        comp.confirmDelete(123);
        tick();

        // THEN
        expect(service.delete).toHaveBeenCalledWith(123);
        expect(mockActiveModal.close).toHaveBeenCalledWith('deleted');
      })
    ));

    it('Should not call delete service on clear', () => {
      // GIVEN
      jest.spyOn(service, 'delete');

      // WHEN
      comp.cancel();

      // THEN
      expect(service.delete).not.toHaveBeenCalled();
      expect(mockActiveModal.close).not.toHaveBeenCalled();
      expect(mockActiveModal.dismiss).toHaveBeenCalled();
    });
  });
});
